Adds default package.json for client
Without this `package.json` file the client cannot be built/started with
the `npm run dev` command.

The package.json comes from
https://github.com/sveltejs/component-template which is the same place
as other setup files for the client.

This file was missing because of the `package.json` entry in the
`.gitignore` file at the root of the repo. I've changed this to ignore
the `package.json` for the root project and to include the
`package.json` for the client.